Trait dock_staking_rewards::pallet::Config

pub trait Config: Config + Config {
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event> + Into<<Self as Config>::Event>;
    type PostUpgradeHighRateDuration: Get<Option<DurationInEras>>;
    type LowRateRewardDecayPct: Get<Percent>;
    type HighRateRewardDecayPct: Get<Percent>;
    type TreasuryRewardsPct: Get<Percent>;
    type RewardCurve: Get<&'static PiecewiseLinear<'static>>;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event> + Into<<Self as Config>::Event>

The overarching event type.

type PostUpgradeHighRateDuration: Get<Option<DurationInEras>>

Optional duration in eras for high-rate rewards to be paid after the upgrade. High-rate rewards will become active at the beginning of the next sequential era.

type LowRateRewardDecayPct: Get<Percent>

The percentage by which remaining emission supply decreases

type HighRateRewardDecayPct: Get<Percent>

High rate percentage by which remaining emission supply decreases. Only used during PostUpgradeHighRateDuration.

type TreasuryRewardsPct: Get<Percent>

The percentage of rewards going to treasury

type RewardCurve: Get<&'static PiecewiseLinear<'static>>

The NPoS reward curve where the first 2 points (of points field) correspond to the lowest and highest inflation and the subsequent points correspond to decreasing inflation